Fixed pagination; Added statusbar; Few view changes; Fixed date and time;

+++ src/main/java/com/fm/internal/services/implementation/PaginationServiceImpl.java
@@ -5,7 +5,7 @@ import com.fm.internal.dtos.PaginationDto;
 public class PaginationServiceImpl {
     public PaginationDto createPagination(long itemId, int selectedPage, int pageSize, long collectionSize, String url) {
         int firstItem = (selectedPage - 1) * pageSize;
-        int pageCount = (int) Math.ceil(collectionSize / pageSize);
+        int pageCount = (int) Math.ceil((double) collectionSize / pageSize);
         int startPage = selectedPage - 5 > 0 ? selectedPage - 5 : 1;
         int endPage = startPage + 10;
         if (endPage > pageCount) {
